+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:50929/","mongoDBName":"jest"}
+{"mongoUri":"mongodb://127.0.0.1:50995/","mongoDBName":"jest"}
