+++ src/main/factories/login/login-validation-factory.ts
@@ -1,6 +1,6 @@
+import { EmailValidatorAdapter } from '@main/adapters/validators/email-validator-adapter'
 import { EmailValidation, RequiredFieldValidation, ValidationComposite } from '@presentation/helpers/validators'
 import { Validation } from '@presentation/protocols/validation'
-import { EmailValidatorAdapter } from '@presentation/utils/email-validator-adapter'
 
 export const makeLoginValidation = (): ValidationComposite => {
   const validations: Validation[] = []
